+++ vck/src/commonMain/kotlin/at/asitplus/wallet/lib/agent/StatusListIssuer.kt
@@ -5,16 +5,20 @@ import at.asitplus.wallet.lib.data.rfc.tokenStatusList.agents.StatusIssuer
 import at.asitplus.wallet.lib.data.rfc.tokenStatusList.agents.StatusProvider
 
 /**
- * Summarizes operations for an Issuer in the sense of the [W3C VC Data Model](https://w3c.github.io/vc-data-model/).
+ * Summarizes operations for a Status Issuer in the sense of
+ * [Token Status List (TSL)](https://www.ietf.org/archive/id/draft-ietf-oauth-status-list-21.html).
  *
- * It can issue Verifiable Credentials, revoke credentials and build a revocation list.
+ * See also [at.asitplus.wallet.lib.data.rfc.tokenStatusList.primitives.TokenStatus].
+ *
+ * It manages the Status List (which describe status of referenced tokens, mostly verifiable credentials),
+ * and issue Status List Tokens (which embed the status list).
  */
 interface StatusListIssuer : StatusIssuer, StatusProvider {
 
     /**
-     * Returns a revocation list which can either be
-     * status list as defined in [TokenListStatus](https://www.ietf.org/archive/id/draft-ietf-oauth-status-list-17.html)
-     * or an identifier list as defined in ISO18013-5
+     * Returns a revocation list which can either be status list as defined in
+     * [Token Status List](https://www.ietf.org/archive/id/draft-ietf-oauth-status-list-21.html)
+     * or an identifier list as defined in ISO/IEC 18013-5:2021.
      */
     fun buildRevocationList(
         timePeriod: Int? = null,
@@ -23,13 +27,15 @@ interface StatusListIssuer : StatusIssuer, StatusProvider {
 
 
     /**
-     * Sets the status of one specific credential to [at.asitplus.wallet.lib.data.rfc.tokenStatusList.primitives.TokenStatus.Invalid].
+     * Sets the status of one specific credential to
+     * [at.asitplus.wallet.lib.data.rfc.tokenStatusList.primitives.TokenStatus.Invalid].
      * Returns true if this credential has been revoked.
      */
     fun revokeCredentialByIndex(timePeriod: Int, statusListIndex: ULong): Boolean
 
     /**
-     * Sets the status of one specific credential to [at.asitplus.wallet.lib.data.rfc.tokenStatusList.primitives.TokenStatus.Invalid].
+     * Sets the status of one specific credential to
+     * [at.asitplus.wallet.lib.data.rfc.tokenStatusList.primitives.TokenStatus.Invalid].
      * Returns true if this credential has been revoked.
      */
     fun revokeCredentialByIdentifier(timePeriod: Int, identifier: ByteArray): Boolean